tests/test_operators.py
"""算子单测：ImageLoad, Caption, Quality, Filter, Sink（mock 推理 API）。"""
import pytest


def test_image_load_op_empty_paths():
    pytest.importorskip("daft")
    from ops.image import ImageLoadOp
    from ops.configs import ImageLoadConfig

    op = ImageLoadOp(ImageLoadConfig(path_source=[]))
    df = op.read()
    assert df is not None
    # Daft DataFrame
    if hasattr(df, "column_names"):
        assert "path" in df.column_names
        assert "image" in df.column_names
    out = df.to_pandas() if hasattr(df, "to_pandas") else df
    assert len(out) == 0


def test_image_load_op_from_list(sample_paths):
    pytest.importorskip("daft")
    from ops.image import ImageLoadOp
    from ops.configs import ImageLoadConfig

    op = ImageLoadOp(ImageLoadConfig(path_source=sample_paths))
    df = op.read()
    out = df.to_pandas() if hasattr(df, "to_pandas") else df
    assert len(out) == 2
    assert list(out["path"]) == sample_paths


def test_caption_inference_op_mock(inference_api_config):
    pytest.importorskip("daft")
    import daft
    from unittest.mock import patch
    from ops.image import CaptionInferenceOp
    from ops.configs import CaptionInferenceConfig

    cfg = CaptionInferenceConfig(prompt="describe", inference_api_config=inference_api_config)
    op = CaptionInferenceOp(cfg)
    df = daft.from_pydict({"path": ["a"], "image": ["placeholder"]})

    with patch("ops.image.caption.call_caption_api", return_value="a cat"):
        out = op._transform(df)
    # 可能走 UDF 分支，需 collect 才有结果；或走 to_pandas 回退
    if hasattr(out, "column_names"):
        assert "caption" in out.column_names
    else:
        assert "caption" in out.columns


def test_quality_check_op_mock(inference_api_config):
    pytest.importorskip("daft")
    import daft
    from unittest.mock import patch
    from ops.image import QualityCheckOp
    from ops.configs import QualityCheckConfig

    cfg = QualityCheckConfig(
        prompt="yes or no",
        threshold=0.8,
        inference_api_config=inference_api_config,
    )
    op = QualityCheckOp(cfg)
    df = daft.from_pydict({"path": ["a"], "image": ["x"], "caption": ["a cat"]})

    with patch("ops.image.quality.call_quality_api", return_value=(True, 0.9)):
        out = op._transform(df)
    if hasattr(out, "column_names"):
        assert "quality_ok" in out.column_names
        assert "quality_score" in out.column_names


def test_filter_by_quality_op():
    pytest.importorskip("daft")
    import daft
    from ops.image import FilterByQualityOp
    from ops.configs import FilterByQualityConfig

    cfg = FilterByQualityConfig(quality_ok_column="quality_ok")
    op = FilterByQualityOp(cfg)
    df = daft.from_pydict({
        "path": ["a", "b", "c"],
        "quality_ok": [True, False, True],
    })
    out = op._transform(df)
    pdf = out.to_pandas() if hasattr(out, "to_pandas") else out
    assert len(pdf) == 2


def test_qualified_image_sink_op(tmp_path):
    pytest.importorskip("daft")
    import daft
    from ops.image import QualifiedImageSinkOp
    from ops.configs import QualifiedImageSinkConfig

    cfg = QualifiedImageSinkConfig(
        output_prefix=str(tmp_path),
        write_metadata=True,
    )
    op = QualifiedImageSinkOp(cfg)
    # 最小 DataFrame：path + image 占位
    df = daft.from_pydict({
        "path": [str(tmp_path / "f1.jpg")],
        "image": [b"fake"],
    })
    # _sink 会 collect，可能报错若 Daft 不支持；用 pandas 模拟
    try:
        op._sink(df)
    except Exception:
        # 若 collect 失败，用 pandas 构造
        import pandas as pd
        pdf = pd.DataFrame({"path": [str(tmp_path / "f1.jpg")], "image": [b"fake"]})
        op._sink(daft.from_pandas(pdf))
    assert (tmp_path / cfg.metadata_filename).exists() or list(tmp_path.iterdir())
